refactor(contentpublisher): finish nesting ContentPublisher into folders matching house style
Finishes nesting the partially-organised ContentPublisher class library, continuing the folder-by-sub-namespace pattern from #896-#905.
