+++ Class-Libraries/RedditPodcastPoster.ContentPublisher/Publishers/DiscoveryPublisher.cs
@@ -4,9 +4,10 @@ using Amazon.S3;
 using Amazon.S3.Model;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using RedditPodcastPoster.ContentPublisher.Configuration;
 using RedditPodcastPoster.ContentPublisher.Models;
 
-namespace RedditPodcastPoster.ContentPublisher;
+namespace RedditPodcastPoster.ContentPublisher.Publishers;
 
 public class DiscoveryPublisher(
     IAmazonS3 client,
