+++ Class-Libraries/RedditPodcastPoster.ContentPublisher/Publishers/IPeoplePublisher.cs
@@ -1,4 +1,4 @@
-namespace RedditPodcastPoster.ContentPublisher;
+namespace RedditPodcastPoster.ContentPublisher.Publishers;
 
 public interface IPeoplePublisher
 {
