+++ Class-Libraries/RedditPodcastPoster.ContentPublisher/Publishers/LanguagesPublisher.cs
@@ -5,8 +5,9 @@ using Amazon.S3;
 using Amazon.S3.Model;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using RedditPodcastPoster.ContentPublisher.Configuration;
 
-namespace RedditPodcastPoster.ContentPublisher;
+namespace RedditPodcastPoster.ContentPublisher.Publishers;
 
 public class LanguagesPublisher(
     IAmazonS3 client,
