+++ Cloud/Api/Services/DiscoveryResultsService.cs
@@ -1,7 +1,7 @@
 using Api.Dtos;
 using Api.Dtos.Extensions;
 using Microsoft.Extensions.Logging;
-using RedditPodcastPoster.ContentPublisher;
+using RedditPodcastPoster.ContentPublisher.Publishers;
 using RedditPodcastPoster.Discovery.Services;
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
